fix: phpstan
